Move some of Texture2DCreator into Texture2DExtensions
Moved the general stuff and kept the stuff that's specific to PAC
