+++ Assets/Scripts/Extensions/UnityEngine/Texture2DExtensions.cs
@@ -80,5 +80,84 @@ namespace PAC.Extensions
             texture.Apply();
             return texture;
         }
+
+        /// <summary>
+        /// Checks that <paramref name="width"/> and <paramref name="height"/> are both &gt; 0, and throws an <see cref="ArgumentOutOfRangeException"/> otherwise.
+        /// </summary>
+        public static void AssertValidTextureDimensions(int width, int height, string widthParamName, string heightParamName)
+        {
+            if (width <= 0)
+            {
+                throw new ArgumentOutOfRangeException($"{widthParamName} is non-positive: {width}.", widthParamName);
+            }
+            if (height <= 0)
+            {
+                throw new ArgumentOutOfRangeException($"{heightParamName} is non-positive: {height}.", heightParamName);
+            }
+        }
+
+        /// <summary>
+        /// Creates a transparent (alpha 0) <see cref="Texture2D"/> of size <paramref name="width"/> x <paramref name="height"/>.
+        /// </summary>
+        /// <remarks>
+        /// Calls <see cref="Texture2D.Apply()"/> on the returned <see cref="Texture2D"/>.
+        /// </remarks>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="width"/> or <paramref name="height"/> is &lt;= 0.</exception>
+        public static Texture2D Transparent(int width, int height) => Solid(width, height, Color.clear);
+
+        /// <summary>
+        /// Creates a <see cref="Texture2D"/> of size <paramref name="width"/> x <paramref name="height"/>, filled with the given colour.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// <see cref="Solid(int, int, Color32)"/> is faster.
+        /// </para>
+        /// <para>
+        /// Calls <see cref="Texture2D.Apply()"/> on the returned <see cref="Texture2D"/>.
+        /// </para>
+        /// </remarks>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="width"/> or <paramref name="height"/> is &lt;= 0.</exception>
+        public static Texture2D Solid(int width, int height, Color colour)
+        {
+            AssertValidTextureDimensions(width, height, nameof(width), nameof(height));
+
+            Texture2D texture = new Texture2D(width, height);
+
+            Color[] pixels = new Color[width * height];
+            for (int index = 0; index < pixels.Length; index++)
+            {
+                pixels[index] = colour;
+            }
+
+            texture.SetPixels(pixels);
+            return texture.Applied();
+        }
+        /// <summary>
+        /// Creates a <see cref="Texture2D"/> of size <paramref name="width"/> x <paramref name="height"/>, filled with the given colour.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// This is faster than <see cref="Solid(int, int, Color)"/>.
+        /// </para>
+        /// <para>
+        /// Calls <see cref="Texture2D.Apply()"/> on the returned <see cref="Texture2D"/>.
+        /// </para>
+        /// </remarks>
+        /// <exception cref="ArgumentOutOfRangeException"><paramref name="width"/> or <paramref name="height"/> is &lt;= 0.</exception>
+        public static Texture2D Solid(int width, int height, Color32 colour)
+        {
+            AssertValidTextureDimensions(width, height, nameof(width), nameof(height));
+
+            Texture2D texture = new Texture2D(width, height);
+
+            Color32[] pixels = new Color32[width * height];
+            for (int index = 0; index < pixels.Length; index++)
+            {
+                pixels[index] = colour;
+            }
+
+            texture.SetPixels32(pixels);
+            return texture.Applied();
+        }
     }
 }
