+++ Assets/Scripts/Files/File.cs
@@ -10,7 +10,6 @@ using UnityEngine.Events;
 using PAC.Json;
 using System.Runtime.Serialization;
 using PAC.Extensions;
-using PAC.ImageEditing;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Extensions;
@@ -508,7 +507,7 @@ namespace PAC.Files
         /// <summary>
         /// Adds a blank normal layer at the given index.
         /// </summary>
-        public void AddNormalLayer(int index) => AddNormalLayer(Texture2DCreator.Transparent(width, height), index);
+        public void AddNormalLayer(int index) => AddNormalLayer(Texture2DExtensions.Transparent(width, height), index);
         /// <summary>
         /// Adds a normal layer at the given index with the given texture.
         /// </summary>
@@ -717,7 +716,7 @@ namespace PAC.Files
 
             if (!inclusive && (highestLayer == lowestLayer || highestLayer == lowestLayer - 1))
             {
-                return Texture2DCreator.Transparent(width, height);
+                return Texture2DExtensions.Transparent(width, height);
             }
 
             // Get the indices of the highest / lowest visible layers so that we don't waste time rendering some invisible layers.
@@ -744,7 +743,7 @@ namespace PAC.Files
 
             if (highestVisibleLayer == -1)
             {
-                return Texture2DCreator.Transparent(width, height);
+                return Texture2DExtensions.Transparent(width, height);
             }
 
             return RenderLayers(IntRange.InclIncl(highestVisibleLayer, lowestVisibleLayer), frame);
@@ -759,7 +758,7 @@ namespace PAC.Files
         {
             if (layerIndices.None())
             {
-                return Texture2DCreator.Transparent(width, height);
+                return Texture2DExtensions.Transparent(width, height);
             }
 
             Texture2D tex = new Texture2D(width, height);
@@ -783,7 +782,7 @@ namespace PAC.Files
         {
             if (layer == 0 && !inclusive)
             {
-                return Texture2DCreator.Transparent(width, height);
+                return Texture2DExtensions.Transparent(width, height);
             }
             return RenderLayers(0, layer - (inclusive ? 0 : 1), frame);
         }
@@ -796,7 +795,7 @@ namespace PAC.Files
         {
             if (layer == layers.Count - 1 && !inclusive)
             {
-                return Texture2DCreator.Transparent(width, height);
+                return Texture2DExtensions.Transparent(width, height);
             }
             return RenderLayers(layer + (inclusive ? 0 : 1), layers.Count - 1, frame);
         }
