+++ Assets/Scripts/Layers/NormalLayer.cs
@@ -4,6 +4,7 @@ using System.Runtime.Serialization;
 using PAC.Animation;
 using PAC.Colour.Compositing;
 using PAC.DataStructures;
+using PAC.Extensions;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.ImageEditing;
@@ -20,9 +21,9 @@ namespace PAC.Layers
     {
         public override LayerType layerType => LayerType.Normal;
 
-        public NormalLayer(int width, int height) : this("", Texture2DCreator.Transparent(width, height)) { }
+        public NormalLayer(int width, int height) : this("", Texture2DExtensions.Transparent(width, height)) { }
         public NormalLayer(Texture2D texture) : this("", texture) { }
-        public NormalLayer(string name, int width, int height) : this(name, Texture2DCreator.Transparent(width, height)) { }
+        public NormalLayer(string name, int width, int height) : this(name, Texture2DExtensions.Transparent(width, height)) { }
         public NormalLayer(string name, Texture2D texture) : base(name, texture) { }
 
         /// <summary>
@@ -211,7 +212,7 @@ namespace PAC.Layers
 
                 if (keyframe == 0)
                 {
-                    AddKeyFrame(0, Texture2DCreator.Transparent(width, height));
+                    AddKeyFrame(0, Texture2DExtensions.Transparent(width, height));
                 }
 
                 return keyFrame;
@@ -222,7 +223,7 @@ namespace PAC.Layers
         public override void ClearFrames()
         {
             keyFrames = new List<AnimationKeyFrame>();
-            AddKeyFrame(0, Texture2DCreator.Transparent(width, height));
+            AddKeyFrame(0, Texture2DExtensions.Transparent(width, height));
         }
 
         public new class JsonConverter : JsonConversion.JsonConverter<NormalLayer, JsonData.Object>
