+++ Assets/Scripts/Layers/TileLayer.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using PAC.Animation;
 using PAC.DataStructures;
 using PAC.Exceptions;
+using PAC.Extensions;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Extensions;
@@ -36,7 +37,7 @@ namespace PAC.Layers
         // This is to help performance - so that we can rerender the key frames once after all tiles have been changed instead of after each one.
         private bool ignoreOnTilePixelsChanged = false;
 
-        public TileLayer(string name, int width, int height) : base(name, Texture2DCreator.Transparent(width, height)) { }
+        public TileLayer(string name, int width, int height) : base(name, Texture2DExtensions.Transparent(width, height)) { }
 
         /// <summary>
         /// Creates a deep copy of the TileLayer.
@@ -414,7 +415,7 @@ namespace PAC.Layers
 
                 if (keyframe == 0)
                 {
-                    AddKeyFrame(0, Texture2DCreator.Transparent(width, height));
+                    AddKeyFrame(0, Texture2DExtensions.Transparent(width, height));
                 }
 
                 return keyFrame;
@@ -430,7 +431,7 @@ namespace PAC.Layers
             }
 
             keyFrames = new List<AnimationKeyFrame>();
-            AddKeyFrame(0, Texture2DCreator.Transparent(width, height));
+            AddKeyFrame(0, Texture2DExtensions.Transparent(width, height));
         }
 
         /// <summary>
